+++ HISTORY.md
@@ -2,7 +2,11 @@
 
 ## in development:
 * **Notable Changes:**
+    * "turbo by default" - keep scheduling non-animating processes between animation ticks
     * pick a random float by entering an integer with a decimal point into at least one of the "pick random" reporter's input slots
+
+### 2025-01-03
+    * objects, threads: "turbo by default" - keep scheduling non-animating processes between animation ticks 
 
 ### 2025-01-02
 * new dev version